Use troposphere to produce CFN tmpl procedurally
A new package in here `cirrostratus_cfn` outputs CloudFormation YAML when
executed. It more or less produces the same YAML as is hand-written and joined
with the tool `yaml-builder` in LexisNexis, except it's done with Python and is
therefore far more flexible.

I hope that that tool, which we could just call BriteFormation or something like
that, can be made flexible enough that the basics of what needs configured can
be provided by some very lightweight config files. Maybe. Or just use
troposphere directly for each project that needs it, and an individual
CloudFormation template for those that don't.