Consolidate secret parameter definition
Defining a secret takes 3 types which I had spread across 3 files as was done in
the example LexisNexis yaml files. This is unnecessary and is now in one file,
which a function that produces all 3 types.

Had to change "resources" function to "items", as it doesn't strictly return
resources anymore. There's a case to be made for putting secrets in a config
file.